moving config outside internal package
